Revert "Multimedia, iOS: Stop build if no FFmpeg frameworks are found"
This reverts commit cb9c09301ae4edc31d13472095fd4373ecaf452a.

Reason for revert:

This change was merged too early. If this stays in as it is,
this will block all integrations happening in between
35fae658f3792d49bedbbb552b18384644a103e2
and a2669c1d60a4f7bc381e7aae70ba7d1f8daba51a on all platforms. This
change should be merged after the two, possibly while only being
applied to release-builds.

Change-Id: I5e7e6040c3ab685669f9216c71b7b874c01da1bc
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>