Camera, Apple, FFmpeg: Make updateCameraFormat take format parameter
Upcoming patches will make some changes to how format is applied.
The current behavior for the Apple FFmpeg backend is to
update the public-facing cameraFormat property even if we fail
to apply it.

This patch performs some preparation for the refactor, by letting
the camera-format apply logic take a parameter for what format
to apply.

No functional changes.

Pick-to: 6.9 6.8
Task-number: QTBUG-132087
Change-Id: I8235277f25081ec409040f7641383b698b1b95f9
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>