Apply redOrAlphaIndex for shaders
This patch allows us to remove all _a shaders apart nv12_a and nv21_a;
nv12_a and nv21_a are renamed to *_fallback, as the suffix _a is not
representative after this refactoring.
The same logic for normal shaders and fallbacks allows us to write
tests and gets around code duplication.

Pick-to: 6.9 6.8
Change-Id: I3fc2f6a2c158961a228de4dd58fb68affba73303
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>