Disable seeking if seeking to media start fails while loading media
FFmpeg demuxers may not be able to support seeking even if they do not
have the AVFMTCTX_UNSEEKABLE flag set. This can for example happen with
rtsp streams if the server does not support seeking.

This patch attempts to seek to the start of the media, and if this
fails, we define the media as not seekable. This allows applications to
disable UI elements that require seeking immediately as the media is
being loaded, and is used to disable seeking in the demuxer.

Task-number: QTBUG-128908
Task-number: QTBUG-116782
Pick-to: 6.9 6.8
Change-Id: Iedd496c8f8943bde7b0c198947ad29d67d30b937
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>