Add AVFormatContext::start_time to seek position
This fixes an issue where QMediaPlayer emits error when loading media
that has a non-zero start_time. This can typically happen when streaming
media over network.

This fixes parts of the issue in the listed JIRA ticket, but there are
remaining issues related to time synchronization, because media frames
are still time-stamped according to the FFmpeg timeline with a non-zero
start-time. This fix merely prevents seek errors when the media is
loaded.

Task-number: QTBUG-128908
Pick-to: 6.9 6.8
Change-Id: Ia31dcf73aafabb6ec11567331749f94aa9398220
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>